rename BertOptimizationOptions to FusionOptions
