+++ onnxruntime/python/tools/transformers/onnx_model_bert.py
@@ -16,34 +16,17 @@ from fusion_gelu import FusionGelu
 from fusion_fastgelu import FusionFastGelu
 from fusion_biasgelu import FusionBiasGelu
 from fusion_gelu_approximation import FusionGeluApproximation
-from fusion_utils import FusionUtils
+from fusion_options import FusionOptions
 
 logger = getLogger(__name__)
 
 
-class BertOptimizationOptions:
+class BertOptimizationOptions(FusionOptions):
+    """ This class is deprecated
+    """
     def __init__(self, model_type):
-        self.enable_gelu = True
-        self.enable_layer_norm = True
-        self.enable_attention = True
-        self.enable_skip_layer_norm = True
-        self.enable_embed_layer_norm = True
-        self.enable_bias_skip_layer_norm = True
-        self.enable_bias_gelu = True
-        self.enable_gelu_approximation = False
-        self.attention_mask_format = AttentionMaskFormat.AttentionMask
-
-        if model_type == 'gpt2':
-            self.enable_skip_layer_norm = False
-
-    def use_raw_attention_mask(self, use_raw_mask=True):
-        if use_raw_mask:
-            self.attention_mask_format = AttentionMaskFormat.AttentionMask
-        else:
-            self.attention_mask_format = AttentionMaskFormat.MaskIndexEnd
-
-    def disable_attention_mask(self):
-        self.attention_mask_format = AttentionMaskFormat.NoMask
+        logger.warning(f"BertOptimizationOptions is depreciated. Please use FusionOptions instead.")
+        super().__init__(model_type)
 
 
 class BertOnnxModel(OnnxModel):
@@ -264,7 +247,7 @@ class BertOnnxModel(OnnxModel):
         self.clean_graph()
         self.prune_graph()
 
-    def optimize(self, options: BertOptimizationOptions = None, add_dynamic_axes=False):
+    def optimize(self, options: FusionOptions = None, add_dynamic_axes=False):
         if (options is None) or options.enable_layer_norm:
             self.fuse_layer_norm()
